fix delete fcm token en logout/close en iOS

+++ components/MuroScreen.js
@@ -195,29 +195,31 @@ export class MuroScreen extends React.Component {
     const authToken = await AppAsyncStorage.getTokenFromSession();
     var myHeaders = new Headers({ 'X-Auth-Token': authToken, });
 
-    // TODO: DELETE del token en app-server
-    let fcmToken = await AsyncStorage.getItem('fcmToken');
-    var myBody = JSON.stringify({ token: fcmToken, });
+    if (Platform.OS !== 'ios') {
+      // TODO: DELETE del token en app-server
+      let fcmToken = await AsyncStorage.getItem('fcmToken');
+      var myBody = JSON.stringify({ token: fcmToken, });
 
-    fetch(EndPoints.fcm, {
-      method: 'DELETE',
-      headers: myHeaders,
-      body: myBody,
-    })
-      .then((response) => response.json().then(json => {
-        return { data: json, fullResponse: response }
-      }))
-      .then((responseJson) => {
-        console.log('------- DELETE Server FCM ------');
-        AppUtils.printResponseJson(responseJson);
+      fetch(EndPoints.fcm, {
+        method: 'DELETE',
+        headers: myHeaders,
+        body: myBody,
       })
-      .catch((error) => {
-        console.log('------- error DELETE Server FCM ------');
-        console.log(error);
-      });
+        .then((response) => response.json().then(json => {
+          return { data: json, fullResponse: response }
+        }))
+        .then((responseJson) => {
+          console.log('------- DELETE Server FCM ------');
+          AppUtils.printResponseJson(responseJson);
+        })
+        .catch((error) => {
+          console.log('------- error DELETE Server FCM ------');
+          console.log(error);
+        });
 
-    // borro token de firebase
-    firebase.messaging().deleteToken();
+      // borro token de firebase
+      firebase.messaging().deleteToken();
+    }
 
     // borro session en el app-server
     fetch(EndPoints.sessions, {
@@ -238,7 +240,6 @@ export class MuroScreen extends React.Component {
         this.setState({ loadingWallVideos: false });
         AppUtils.logout();
       });
-    
   }
 
   async requestWallVideos() {
