+++ components/ProfileScreen.js
@@ -1,5 +1,6 @@
 /* Import Libs */
 import React from 'react';
+import { Platform } from 'react-native';
 import { StackActions } from '@react-navigation/native';
 import { ScrollView, StyleSheet, Image, Text, View, Alert } from 'react-native';
 import { Button, Divider, Chip, Card, Appbar, ActivityIndicator, Snackbar } from 'react-native-paper';
@@ -232,31 +233,33 @@ export class ProfileScreen extends React.Component {
     const authToken = await AppAsyncStorage.getTokenFromSession();
     var myHeaders = new Headers({ 'X-Auth-Token': authToken, });
 
-    // DELETE del token en app-server
-    let fcmToken = await AsyncStorage.getItem('fcmToken');
-    var myBody = JSON.stringify({ token: fcmToken, });
+    if (Platform.OS !== 'ios') {
+      // DELETE del token en app-server
+      let fcmToken = await AsyncStorage.getItem('fcmToken');
+      var myBody = JSON.stringify({ token: fcmToken, });
 
-    console.log('fcmToken: ' + fcmToken);
+      console.log('fcmToken: ' + fcmToken);
 
-    fetch(EndPoints.fcm, {
-      method: 'DELETE',
-      headers: myHeaders,
-      body: myBody,
-    })
-      .then((response) => response.json().then(json => {
-        return { data: json, fullResponse: response }
-      }))
-      .then((responseJson) => {
-        console.log('------- DELETE Server FCM ------');
-        AppUtils.printResponseJson(responseJson);
+      fetch(EndPoints.fcm, {
+        method: 'DELETE',
+        headers: myHeaders,
+        body: myBody,
       })
-      .catch((error) => {
-        console.log('------- error DELETE Server FCM ------');
-        console.log(error);
-      });
+        .then((response) => response.json().then(json => {
+          return { data: json, fullResponse: response }
+        }))
+        .then((responseJson) => {
+          console.log('------- DELETE Server FCM ------');
+          AppUtils.printResponseJson(responseJson);
+        })
+        .catch((error) => {
+          console.log('------- error DELETE Server FCM ------');
+          console.log(error);
+        });
 
-    // borro token de firebase
-    firebase.messaging().deleteToken();
+      // borro token de firebase
+      firebase.messaging().deleteToken();
+    }
 
     // borro session en el app-server
     fetch(EndPoints.users + "/" + username, {
